BACKPORT: drm/i915: Use a plain old int for the cdclk/mdclk ratio
No point in throwing around u8 when we're dealing with
just an integer. Use a plain old boring 'int'.

Link: https://patchwork.freedesktop.org/patch/msgid/20240402155016.13733-14-ville.syrjala@linux.intel.com
(cherry picked from commit 96c420d5016e5bec57dccb161744c43366b72eca)

Conflicts:
   drivers/gpu/drm/i915/display/intel_cdclk.c

BUG=b:350121902
TEST=No corruption should be seen on 4k@144hz DP when eDP lid is closed

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5787852
Tested-by: Altaf Basha <altaf.basha@intel.com>

+++ drivers/gpu/drm/i915/display/intel_cdclk.c
@@ -1789,8 +1789,8 @@ static bool mdclk_source_is_cdclk_pll(struct drm_i915_private *i915)
 	return DISPLAY_VER(i915) >= 20;
 }
 
-u8 intel_mdclk_cdclk_ratio(struct drm_i915_private *i915,
-			   const struct intel_cdclk_config *cdclk_config)
+int intel_mdclk_cdclk_ratio(struct drm_i915_private *i915,
+			    const struct intel_cdclk_config *cdclk_config)
 {
 	if (mdclk_source_is_cdclk_pll(i915))
 		return DIV_ROUND_UP(cdclk_config->vco, cdclk_config->cdclk);
@@ -3220,7 +3220,7 @@ int intel_modeset_calc_cdclk(struct intel_atomic_state *state)
 
 	if (intel_mdclk_cdclk_ratio(dev_priv, &old_cdclk_state->actual) !=
 	    intel_mdclk_cdclk_ratio(dev_priv, &new_cdclk_state->actual)) {
-		u8 ratio = intel_mdclk_cdclk_ratio(dev_priv, &new_cdclk_state->actual);
+		int ratio = intel_mdclk_cdclk_ratio(dev_priv, &new_cdclk_state->actual);
 
 		ret = intel_dbuf_state_set_mdclk_cdclk_ratio(state, ratio);
 		if (ret)
